Log getMetadata
