+++ crates/pixi-build-backend/src/server.rs
@@ -1,11 +1,13 @@
-use std::{net::SocketAddr, sync::Arc};
+use std::{net::SocketAddr, path::Path, sync::Arc};
 
+use fs_err::tokio as tokio_fs;
 use jsonrpc_core::{serde_json, to_value, Error, IoHandler, Params};
-use miette::{IntoDiagnostic, JSONReportHandler};
+use miette::{Context, IntoDiagnostic, JSONReportHandler};
 use pixi_build_types::procedures::{
     self, conda_build::CondaBuildParams, conda_metadata::CondaMetadataParams,
     initialize::InitializeParams, negotiate_capabilities::NegotiateCapabilitiesParams,
 };
+
 use tokio::sync::RwLock;
 
 use crate::protocol::{Protocol, ProtocolInstantiator};
@@ -105,8 +107,14 @@ impl<T: ProtocolInstantiator> Server<T> {
                 async move {
                     let params: CondaMetadataParams = params.parse()?;
                     let state = state.read().await;
-                    state
-                        .as_endpoint()?
+                    let endpoint = state.as_endpoint()?;
+
+                    let debug_dir = endpoint.debug_dir();
+                    log_conda_get_metadata(debug_dir, &params)
+                        .await
+                        .map_err(convert_error)?;
+
+                    endpoint
                         .conda_get_metadata(params)
                         .await
                         .map(|value| to_value(value).expect("failed to convert to json"))
@@ -150,4 +158,29 @@ fn convert_error(err: miette::Report) -> jsonrpc_core::Error {
         message: err.to_string(),
         data: Some(data),
     }
+}
+
+async fn log_conda_get_metadata(
+    debug_dir: Option<&Path>,
+    params: &CondaMetadataParams,
+) -> miette::Result<()> {
+    let Some(debug_dir) = debug_dir else {
+        return Ok(());
+    };
+
+    let json = serde_json::to_string_pretty(&params)
+        .into_diagnostic()
+        .context("failed to serialize parameters to JSON")?;
+
+    tokio_fs::create_dir_all(&debug_dir)
+        .await
+        .into_diagnostic()
+        .context("failed to create data directory")?;
+
+    let path = debug_dir.join("conda_metadata_params.json");
+    tokio_fs::write(&path, json)
+        .await
+        .into_diagnostic()
+        .context("failed to write JSON to file")?;
+    Ok(())
 }
