+++ crates/pixi-build-python/src/protocol.rs
@@ -1,4 +1,4 @@
-use std::{path::PathBuf, str::FromStr, sync::Arc};
+use std::{path::Path, str::FromStr, sync::Arc};
 
 use miette::{Context, IntoDiagnostic};
 use pixi_build_backend::{
@@ -38,8 +38,8 @@ use crate::{
 
 #[async_trait::async_trait]
 impl<P: ProjectModel + Sync> Protocol for PythonBuildBackend<P> {
-    fn debug_dir(&self) -> Option<PathBuf> {
-        self.config.debug_dir.clone()
+    fn debug_dir(&self) -> Option<&Path> {
+        self.config.debug_dir.as_deref()
     }
 
     async fn conda_get_metadata(
